Refactor Searchable into Homebrew::Search
