+++ Library/Homebrew/search.rbi
@@ -1,0 +1,5 @@
+# typed: strict
+
+module Homebrew::Search
+  include Kernel
+end
